+++ internal/execext/exec.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"io"
 	"os"
+	"path/filepath"
 	"strings"
 
 	"mvdan.cc/sh/expand"
@@ -72,6 +73,7 @@ func IsExitError(err error) bool {
 // Expand is a helper to mvdan.cc/shell.Fields that returns the first field
 // if available.
 func Expand(s string) (string, error) {
+	s = filepath.ToSlash(s)
 	s = strings.Replace(s, " ", `\ `, -1)
 	fields, err := shell.Fields(s, nil)
 	if err != nil {